+++ apps/BFF/src/modules/auth/auth.module.ts
@@ -1,12 +1,13 @@
 import { Module } from '@nestjs/common';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
+import { GoogleAuthService } from './google-auth.service';
 import { AppConfigModule } from '../config/config.module';
 
 @Module({
   imports: [AppConfigModule],
   controllers: [AuthController],
-  providers: [AuthService],
+  providers: [AuthService, GoogleAuthService],
   exports: [AuthService],
 })
 export class AuthModule {}
